+++ paper/paper.bib
@@ -37,6 +37,7 @@
   month = jul,
   year = {2018},
   url = {https://arxiv.org/abs/1703.03512},
+  doi={10.1109/LRA.2018.2795645},
   month_numeric = {7}
 }
 
@@ -74,6 +75,7 @@
   timestamp = {2019-10-17T14:19:43.000+0200},
   title = {MuJoCo: A physics engine for model-based control.},
   url = {http://dblp.uni-trier.de/db/conf/iros/iros2012.html#TodorovET12},
+  doi={10.1109/iros.2012.6386109},
   year = 2012
 }
 
