### 2020-07-10 #2459 @vincentberenz
@kyleniemeyer @gkthiruvathukal
Thanks a lot for starting all this. First submission to JOSS, so looking forward to see how the review goes !
