### 2020-07-10 2aa8fab
fixed typo in readme
