+++ src/time.cpp
@@ -20,4 +20,4 @@ long int time_diff_us(const TimePoint& before, const TimePoint& after)
 {
     return std::chrono::duration_cast<Microseconds>(after - before).count();
 }
-}
+}  // namespace o80
