+++ include/o80/void_state.hpp
@@ -60,4 +60,4 @@ public:
     {
     }
 };
-}
+}  // namespace o80
