### 2020-08-10 #2459 @SteveMacenski
I don't think I'm suited to this paper, I'm a robotics expert.
