### 2020-08-11 #2459 @gkthiruvathukal
@giodegas If you scroll up, you can see the "Check article proof" link.

@SteveMacenski Thank you for responding.
