### 2020-08-11 #2459 @gkthiruvathukal
@giodegas You had asked about the review process and criteria. Here are some helpful links from our journal docs:

Here are the reviewer guidelines:
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

And here is a bit more detail about the review criteria:
https://joss.readthedocs.io/en/latest/review_criteria.html
