### 2020-08-11 #2459 @gkthiruvathukal
@traversaro, Are you willing to contribute a review for this JOSS submission? I need two reviewers before I can proceed.
