### 2020-08-12 6c03dba
templated config for python bindings
