+++ include/o80/pybind11_helper.hpp
@@ -18,42 +18,15 @@
 
 namespace o80
 {
-/* ! Configuration for the creating of python bindings.
- *   By default, bindings for these (templated) classes will
- *   binded: States, State, Observation, ExtendedState,
- *   FrontEnd and Backend. This class allows to inhibite the
- *   creation of bindings on selected classes, by setting the
- *   related bool value to false.
- *   This class also allows to add a prefix to the name of the
- *   binded class. For example, if the prefix is "MyRobot",
- *   the name of the FrontEnd binded class will be "MyRobotFrontEnd".
- */
-class Pybind11Config
-{
-public:
-    Pybind11Config(bool all_false = false);
-    bool states = true;
-    bool state = true;
-    bool observation = true;
-    bool extended_state = true;
-    bool frontend = true;
-    bool backend = true;
-    std::string prefix;
-};
 
-/**
- * ! Creates in python module m python bindings
- * for the classes States, State, ExtendedState, Observation,
- * FrontEnd and Backend, except if inhibited by the
- * the instance of Pybind11Config. The name of the created
- * class will have the prefix specified by the config.
- */
-template <int QUEUE_SIZE,
-          int NB_ACTUATORS,
-          class o80_STATE,
-          class o80_EXTENDED_STATE>
-void create_core_python_bindings(pybind11::module &m,
-                                 Pybind11Config pybind11_config);
+  typedef std::integral_constant<int,0> NO_STATES;
+  typedef std::integral_constant<int,1> NO_STATE;
+  typedef std::integral_constant<int,2> NO_OBSERVATION;
+  typedef std::integral_constant<int,3> NO_EXTENDED_STATE;
+  typedef std::integral_constant<int,4> NO_FRONTEND;
+  typedef std::integral_constant<int,5> NO_BACKEND;
+
+
 
 /**
  * ! Creates in python module m python bindings
@@ -63,19 +36,16 @@ void create_core_python_bindings(pybind11::module &m,
 template <int QUEUE_SIZE,
           int NB_ACTUATORS,
           class o80_STATE,
-          class o80_EXTENDED_STATE>
-void create_core_python_bindings(pybind11::module &m);
-
-template <int QUEUE_SIZE,
-          int NB_ACTUATORS,
-          class o80_STATE,
-          class RobotDriver,
-          class RobotStandalone,
           class o80_EXTENDED_STATE,
-          typename... DriverArgs>
-void _create_python_bindings(pybind11::module &m,
-                             Pybind11Config pybind11_config);
+	  typename ... EXCLUDED_CLASSES>
+void create_python_bindings(pybind11::module &m,
+				 std::string prefix=std::string(""));
 
+template <class RobotStandalone,
+	  typename ... EXCLUDED_CLASSES>
+void create_python_bindings(pybind11::module &m,
+				 std::string prefix=std::string(""));
+  
 /**
  * ! Creates the python bindings for the classes
  *   States, State, ExtendedState, Observation,
@@ -87,8 +57,8 @@ void _create_python_bindings(pybind11::module &m,
  *                        RobotDriver.
  */
 template <class RobotDriver, class RobotStandalone, typename... DriverArgs>
-void create_python_bindings(pybind11::module &m,
-                            Pybind11Config pybind11_config);
+void create_standalone_python_bindings(pybind11::module &m,
+				       std::string prefix=std::string(""));
 
 #include "pybind11_helper.hxx"
 }  // namespace o80
