+++ include/o80/state2d.hpp
@@ -4,15 +4,5 @@
 
 namespace o80
 {
-/* ! StateXd with 2 double attributes */
-class State2d : public StateXd<State2d, double, double>
-{
-public:
-    State2d(double x, double y) : StateXd<State2d, double, double>(x, y)
-    {
-    }
-    State2d() : StateXd<State2d, double, double>()
-    {
-    }
-};
+  typedef StateXd<double, double> State2d;
 }  // namespace o80
