### 2020-08-21 ee92ec4
supporting sensor state (i.e. states that can not be interpolated)
