+++ include/o80/state3d.hpp
@@ -4,5 +4,5 @@
 
 namespace o80
 {
-  typedef StateXd<double,double,double> State3d;
+typedef StateXd<double, double, double> State3d;
 }  // namespace o80
