+++ include/o80/sensor_state.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace o80
+{
+class SensorState
+{
+};
+
+}  // namespace o80
