+++ include/o80/back_end.hpp
@@ -3,10 +3,12 @@
 
 #pragma once
 
+#include <type_traits>
 #include "o80/frequency_measure.hpp"
 #include "o80/logger.hpp"
 #include "o80/memory_clearing.hpp"
 #include "o80/observation.hpp"
+#include "o80/sensor_state.hpp"
 #include "o80/states.hpp"
 #include "o80_internal/controllers_manager.hpp"
 #include "time_series/multiprocess_time_series.hpp"
