### 2020-09-02 f0a5c8b
using time series factories instead of constructors
