### 2020-09-04 00213cd
removed deprecated python wrappers over log
