### 2020-09-08 abbdc0d
added pulse_prepare_wait, unit tests lacking
