+++ include/o80/front_end.hpp
@@ -167,10 +167,26 @@ public:
 
     /*! write all buffered commands to the multiprocess time series commands
      *  (i.e. the related backend will read and execute them), then return
-     * the latest observation/
+     * the latest observation once all commands have been completed
      */
     Observation<NB_ACTUATORS, ROBOT_STATE, EXTENDED_STATE> pulse_and_wait();
 
+    /*! write all buffered commands to the multiprocess time series commands
+     *  (i.e. the related backend will read and execute them), then return
+     * the latest observation. Optionaly, this call may be followed by a 
+     * call to wait, which will be blocking until completion of all commands
+     */
+    Observation<NB_ACTUATORS, ROBOT_STATE, EXTENDED_STATE> pulse_prepare_wait();
+
+  /*! If the previous call to this instance of frontend was a call to
+   * "pulse_prepare_wait", a call to wait will be blocking until
+   * all commands added to the multiprocess time series commands have
+   * been executed. If following call to another of the "pulse" method,
+   * thows a runtime_error.
+   */
+  Observation<NB_ACTUATORS, ROBOT_STATE, EXTENDED_STATE> wait();
+
+  
     /*! write all buffered commands to the multiprocess time series commands
      *  (i.e. the related backend will read and execute them), then wait for the
      *  backend to finish executation of all commands, then return the latest
@@ -218,6 +234,11 @@ private:
 
     // in burst mode: used to the send the activating signal to the backend.
     LeaderPtr leader_;
+
+  // for the use of prepare_wait
+  int completed_index_;
+  bool wait_prepared_;
+  
 };
 
 #include "front_end.hxx"
