### 2020-09-30 b0eeee6
fixed bug of blocking pulse and wait
