### 2020-10-06 40eefb3
frequency manager returns wait time
