### 2020-10-15 #2459 @traversaro
> @traversaro, Are you willing to contribute a review for this JOSS submission? I need two reviewers before I can proceed.

Sorry for the delay, I will be happy to review this submission.
