### 2020-10-15 #2752 @whedon
[REVIEW]: The o80 C++ templated toolbox: Designing customized Python APIs for synchronizing realtime processes
**Submitting author:** @vincentberenz (<a href="http://orcid.org/0000-0003-2226-9660">Vincent Berenz</a>)
**Repository:** <a href="https://github.com/intelligent-soft-robots/o80" target ="_blank">https://github.com/intelligent-soft-robots/o80</a>
**Version:** v1.0
