### 2020-10-15 #2752 @traversaro
Hi @vincentberenz , I see that you submitted the version o80 v1.0 for review, but I can't find it in the repo any reference to any v1.0 version, can you clarify this? Thanks!
