### 2020-10-15 #2752 @vincentberenz
Hi @traversaro , thanks a lot for accepting to review the paper. My apologies, I missed that a tagged/version had to set for review. The master branch is meant to always be the latest stable version. If needed, I'd be happy to tag the latest commit as version 1.0.
