### 2020-10-15 intelligent-soft-robots/o80#2 @traversaro
Provide detailed installation instructions
It may be convenient for users to have easily accessible and usable installation instructions, ideally easily accessible from the repo README or the main documentation page (the current installation docs takes 4 clicks to reach from the README, assuming that you know where it is).

There is currently the documentation in https://intelligent-soft-robots.github.io/code_documentation/o80/docs/html/doc/02.installation.html , but it is incomplete, in particular in the following points:

> installing a few dependencies (apt and pip installation only)

The required apt and pip packages should be clearly listed, ideally by providing the exact commands that can be copy-pasted in the terminal to install them. I just now noticed that the dependencies are indeed listed in https://intelligent-soft-robots.github.io/code_documentation/o80_example/docs/html/doc/01_installation.html, but this is not easily reachable and in general at least the install documentation in a project should be self contained.
