### 2020-10-15 #2752 @traversaro
>
>
> Hi @traversaro , thanks a lot for accepting to review the paper. My apologies, I missed that a tagged/version had to set for review. The master branch is meant to always be the latest stable version. If needed, I'd be happy to tag the latest commit as version 1.0.

Besides the specific requirement of the JOSS review, it seems to be that at least the repos https://github.com/intelligent-soft-robots/o80 are https://github.com/intelligent-soft-robots/o80_example are relatively tightly coupled (one is the software and one is the canonical example on how to use it). Unless you ensure perpetual API stability (that I could not see in the docs) as a user if I start developing against a given version of o80 it would be useful to know what is the corresponding version of https://github.com/intelligent-soft-robots/o80_example . This information probably can be obtained looking at the commit dates, but having a clear version to couple the two could be convenient for the user.
