### 2020-10-15 #2752 @traversaro
Related to installation instructions, I opened an issue in the repo: https://github.com/intelligent-soft-robots/o80/issues/2 .
