### 2020-10-15 #2752 @gkthiruvathukal
@traversaro  I think we're ok here and that review can proceed, right?  I think everything being reviewed should be tagged v1.0. You raise good points about the coupled example. Even though that might not show up in the submission metadata, it should be tagged v1.0, too. I think it is safe to proceed with review nevertheless. At the end, based on all issues raised, the authors will probably make some changes and tag everything again at the end. Hope that helps!
