### 2020-10-15 #2752 @traversaro
Thanks, that is clear to me. For sure the review can continue and is not blocked by this, I just wanted to clarify things.
