### 2020-10-16 #2752 @vincentberenz
@traversaro @gkthiruvathukal I tagged both o80 and o80_example with version 1.0. Thanks for the comments on the alignment of versions between o80 and o80_example, as well as for the issue related to the installation procedure and the broken links.
This is my first submission to JOSS: what is the recommended way ? Answering comments/issues as they are arise, in an interactive fashion, or wait for the review to be done before addressing the issues ?
