### 2020-10-16 #2752 @traversaro
>  This is my first submission to JOSS: what is the recommended way ? Answering comments/issues as they are arise, in an interactive fashion, or wait for the review to be done before addressing the issues ?

It is the first review of JOSS as well. I don't know if there is any recommendation on that, but looking at the other reviews, it seems that an interactive review seems to be possible. Personally I think that interactive review would be desirable, for example because I wanted to test the example, and I already found problems in following the docs for installation, so for me it would be ideal if the installation docs could be fixed, so that I could then try the example following the new installation docs.
