### 2020-10-19 6f584af
added link to installation page to README, add explicit list of apt and pip dependencies to 02_installation
