### 2020-10-20 intelligent-soft-robots/o80#2 @vincentberenz
@traversaro

thanks for finding these issues

- the lists of dependencies has been added to https://intelligent-soft-robots.github.io/code_documentation/o80/docs/html/doc/02.installation.html
- the [readme](https://github.com/intelligent-soft-robots/o80) of o80 has now a direct link to the installation instruction page
- the broken links have been fixed
