### 2020-10-20 intelligent-soft-robots/o80#2 @traversaro
Hi @vincentberenz , thanks for handling this. However, to be honest it is still not trivial to me how to install o80 and try some examples. In particular, I have no idea how (I actually have, but I tried to be a naive user for the sake of providing useful feedback) I am supposed to to the following steps:
> creating a catkin workspace with o80 + dependencies packages (we use the project manager Treep to make this trivial)

Even if not completely self-contained, I would expect to have some pointers to the docs on how to install catkin on Ubuntu 18.04, and if necessary on how to install Treep. Furthermore, what are the `+ dependencies packages` ? If there is some more detailed documentation on some of the example packages, then I think it could make it sense to link it also from the o80 documentation.
