### 2020-10-20 intelligent-soft-robots/o80#2 @vincentberenz
@traversaro looking back at this, I can see this is a valid concern. Here what I propose to do:

- the more detailed and practical installation procedure is provided in the o80_example. I believe this makes sense, as o80_example is the package that will result into concrete implementations of classes (and demos one can run). But the invitation to the user to look at the o80_example installation procedure is given late and is not emphasized. I can improve this.

- I can also improve the installation procedure of o80_example by providing better explanation regarding treep and catkin, as you suggest

- On top on this I can provide a docker image with o80_example already installed and ready to run
