### 2020-10-20 intelligent-soft-robots/o80#2 @traversaro
Great @vincentberenz, I agree on all points.
