### 2020-10-21 bd9114c
documentation installation procedure redirects more clearly to o80_example
