### 2020-10-21 intelligent-soft-robots/o80#2 @vincentberenz
@traversaro I updated the docs
