### 2020-10-25 #2752 @traversaro
Hi @gkthiruvathukal, I provided some preliminary comments to @vincentberenz on the installation procedure, that @vincentberenz addressed, now I need to install the software using the new docs, and play a bit with the examples to go forward with the review.
