### 2020-10-26 intelligent-soft-robots/o80#2 @vincentberenz
@traversaro I added instructions on how to run the demos through docker containers
