### 2020-10-27 611ad49
converted package and CMakeLists to colcon, not tested yet
