+++ CMakeLists.txt
@@ -1,45 +1,41 @@
-cmake_minimum_required(VERSION 2.8.7)
+cmake_minimum_required(VERSION 3.10.2)
 
 project(o80)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED on)
 
-find_package(catkin REQUIRED COMPONENTS
-  pybind11_catkin
+################
+# Dependencies #
+################
+
+find_package(ament_cmake REQUIRED)
+find_package(ament_cmake_python REQUIRED)
+find_package(mpi_cmake_modules REQUIRED)
+find_package(pybind11_vendor REQUIRED)
+find_package(pybind11 REQUIRED)
+find_package(synchronizer REQUIRED)
+find_package(shared_memory REQUIRED)
+find_package(real_time_tools REQUIRED)
+find_package(time_series REQUIRED)
+ament_export_dependencies(
+  mpi_cmake_modules
+  pybind11
   synchronizer
   shared_memory
   real_time_tools
-  time_series)
+  time_series )
 
-# defined in catkin_package mpi_cmake_modules,
-# used to detected for example if using real time computer,
-# also host various cmake files
-DEFINE_OS()
-
-catkin_package(
- LIBRARIES  o80
- INCLUDE_DIRS include internal
- CATKIN_DEPENDS
- pybind11_catkin
- shared_memory
- real_time_tools
- synchronizer
- time_series
-)
-
-include_directories(
-  internal
-  include
-  ${catkin_INCLUDE_DIRS}
-)
+# prepare to export all needed targets
+set(all_targets)
+set(all_target_exports)
 
 
-################
-# Main Library #
-################
+###########
+# Library #
+###########
 
-add_library( o80
+add_library( ${PROJECT_NAME} SHARED
   src/time_stamp.cpp
   src/time.cpp
   src/command_type.cpp
@@ -48,34 +44,68 @@ add_library( o80
   src/void_state.cpp
   src/frequency_manager.cpp
   src/frequency_measure.cpp )
-target_link_libraries(o80 ${catkin_LIBRARIES})
+target_include_directories(
+  ${PROJECT_NAME} PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
+  $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/internal>
+  $<INSTALL_INTERFACE:include>)
+ament_target_dependencies(${PROJECT_NAME}
+  synchronizer
+  shared_memory
+  real_time_tools
+  time_series )
+ament_export_targets(export_${PROJECT_NAME} HAS_LIBRARY_TARGET)
+list(APPEND all_targets ${PROJECT_NAME})
+list(APPEND all_target_exports export_${PROJECT_NAME})
+
 
 ###################
-# Python bindings #
+# Python wrappers #
 ###################
 
-pybind11_add_module(o80_py srcpy/wrappers.cpp)
-target_link_libraries(
-  o80_py PRIVATE o80 ${catkin_LIBRARIES})
-set_target_properties(o80_py PROPERTIES
-  LIBRARY_OUTPUT_DIRECTORY ${CATKIN_DEVEL_PREFIX}/${CATKIN_GLOBAL_PYTHON_DESTINATION}
-  OUTPUT_NAME o80
-  )
-install(TARGETS o80_py DESTINATION ${CATKIN_GLOBAL_PYTHON_DESTINATION})
+add_library(${PROJECT_NAME}_py MODULE srcpy/wrappers.cpp)
+target_link_libraries(${PROJECT_NAME}_py PRIVATE pybind11::module)
[…]
