### 2020-10-29 #2459 @vincentberenz
@traversaro @giodegas

I am working on generating a debian for o80. I will finalize this in the upcoming days (i.e. broader support, documentation, etc), but I though sharing it with you early on may help with the review.

https://owncloud.tuebingen.mpg.de/index.php/s/saSsfK6G5G7aAj9

```bash
sudo apt install ./o80_1.00-1.deb
```

limitations:
- only amd64
- assumes default python3 is python3.6.*

I tested on both my desktop (ubuntu 18.04) and a vanilla ubuntu 18.04 docker image.

This may require fyplot:

```bash
sudo pip3 install fyplot
```

To sanity check installation worked, in a python3 terminal:

```python
import o80
```

if this works (thanks for letting me know of any difficulty), you should be able to run the demos:

```bash
git clone https://github.com/intelligent-soft-robots/o80_example.git
cd o80_example/demos
```

in a first terminal:

```bash
python3 ./standalone_backend.py
# note: has to be python3
```

in a second terminal:

```bash
python3 ./duration_commands.py
# or another demo
```

If first there is another configuration you'd like me to test this on (e.g. other version of ubuntu), please let me know
