### 2020-11-01 345b0e2
BoolState derives from SensorState
