### 2020-11-25 #2752 @gkthiruvathukal
Apologies for the delay in getting a replacement reviewer.

@davidbeckingsale or @vissarion Would you be willing to contribute a review for this JOSS submission?
