### 2020-12-04 #2752 @davidbeckingsale
> Apologies for the delay in getting a replacement reviewer.
>
> @davidbeckingsale or @vissarion Would you be willing to contribute a review for this JOSS submission?

I don't have the bandwidth right now. Thanks for asking.
