### 2020-12-22 intelligent-soft-robots/o80#3 @vissarion
Broken links in README
In the README the button links for building, documentation and tests CI are broken.
e.g. https://github.com/intelligent-soft-robots/o80/blob/master/url
