### 2021-01-18 #2752 @traversaro
>
>
> @traversaro and @vissarion, happy new year! Can. you let me know where you are with respect to the review for this JOSS submission?

Hi @gkthiruvathukal ! Thanks for the ping. I plan to resume the review this week.
