### 2021-01-18 #2752 @vissarion
@gkthiruvathukal happy new year! I paused my review waiting for a feedback in https://github.com/intelligent-soft-robots/o80/issues/3 since it seems there are some works in progress on the package.
