### 2021-01-19 #2752 @vincentberenz
@gkthiruvathukal thanks for pinging this !
@vissarion @traversaro  : indeed, I am working on the packaging. For the build system, we are moving from catkin (used by ROS) to colcon/ament (used by ROS2). In parallel, I am creating a ppa to make installation of the compiled software trivial. All this is taking much more time and effort than expected. And also not as smooth as it should, you may notice the tag for compilation, unit tests and doc building are all red ... Obviously, this make reviewing much harder than it should (but it can wait, if ok with you). My apologies, switch to colcon and setup of a ppa were not part of the plan yet when I requested the review. Note that on the other hand the source code is not expected to change (beyond possible bug fixes).
