### 2021-02-03 faea0f1
added documentation of observation get_time_stamp
