+++ README.md
@@ -1,7 +1,7 @@
 # o80
 Synchronization of c++ processes via a custom python API.
 
-See : the [documentation](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/index.html), an [example project](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/index.html) and the [installation instructions](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html)
+See : the [documentation](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/index.html), an [example project](https://github.com/intelligent-soft-robots/o80_example) and the [installation instructions](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html)
 
 
 # Author
