+++ doc/02.installation.md
@@ -1,38 +1,120 @@
 # Installation
 
-o80 is a c++ library with python wrappers. It uses catkin for compilation.
-It has been tested only on ubuntu 18.04 and python3.
+o80 has been tested exclusively on ubuntu 20.04 and ubuntu 18.04.
 
-o80 is mostly a templated header library. It can thus only be installed in relation with a package which provides concrete classes to be used as template. Thus the detailed installation procedure (that ends up with demos you will be able to run) is available along with a package providing such concrete classees:  [o80 example](https://github.com/intelligent-soft-robots/o80_example).
+## from binaries
 
- **For detailed installation procedure, we invite the reader to refer to the documentation of [o80_example](https://intelligent-soft-robots.github.io/code_documentation/o80_example/docs/html/index.html)**. 
+o80's binaries are available only for ubuntu18.04/python3.6 or ubuntu20.04/python3.8.
 
-The rest of this page provides only a general outline of installation procedure.
+To install:
 
+First, select a version by visiting : visit http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/ or http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/
 
-To use o80, you will need:
+Then, for example:
 
-- a machine installed with ubuntu 18.04, possibly with a preempt patch kernel, if your robot requires it.
-- installing a few dependencies (apt and pip installation only):
 ```bash
-apt install -y ssh \
-               python3-pip \
-               git \
-	       cmake \
-	       libcereal-dev \
-	       libboost-all-dev \
-	       libgtest-dev \
-	       libeigen3-dev \
-	       libedit-dev \
-	       libncurses5-dev \
-	       freeglut3-dev \
-	       libxmu-dev
-	       
-pip3 install catkin_tools treep fyplot pyside2 empy
+wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_ubuntu20.04_py3.8_1.0.tar.gz
+tar -zxvf ./o80_ubuntu20.04_py3.8_1.0.tar.gz
+sudo ./apt_dependencies
+sudo ./pip3_dependencies
+./configure
+sudo make install
+sudo ldconfig
 ```
-- creating a catkin workspace with o80 + dependencies packages (we use the project manager [Treep](https://pypi.org/project/treep/) to make this trivial)
-- creating a catkin project that will host your c++ code. Your c++ code will be the classes used for State, Observation, as well as the Driver interfacing
-  o80 and your robot (see [next page](https://intelligent-soft-robots.github.io/code_documentation/o80/docs/html/doc/03.integration.html) for explanation of what State, Observation and Driver are). o80 has been designed such that this code may be minimalist.
-- creating the python binders (o80 provides some high level functions to makes this easier)
+
+## from source
+
+### from tar
+
+Select a version by visiting : visit http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/ or http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/
+
+Then, for example:
+
+```bash
+wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_source.tar.gz
+tar -zxvf ./o80_ubuntu20.04_py3.8_1.0.tar.gz
+sudo ./apt_dependencies
+sudo ./pip3_dependencies
+./configure
+make
+sudo make install
+sudo ldconfig
+```
 
 
+### via colcon workspace
+
+[Colcon](https://colcon.readthedocs.io/en/released/) is the built system of [ROS2](https://docs.ros.org/en/foxy/index.html).
[…]
