### 2021-03-24 b67d5a9
improved installation doc
