+++ doc/02.installation.md
@@ -2,15 +2,11 @@
 
 o80 has been tested exclusively on ubuntu 20.04 and ubuntu 18.04.
 
-## from binaries
+## From binaries
 
-o80's binaries are available only for ubuntu18.04/python3.6 or ubuntu20.04/python3.8.
+o80's binaries are available only for ubuntu18.04/python3.6 and ubuntu20.04/python3.8.
 
-To install:
-
-First, select a version by visiting : visit http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/ or http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/
-
-Then, for example:
+To install, first select a version by visiting : [http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/](http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/) or [http://people.tuebingen.mpg.de/mpi-is-software/o80/older/](http://people.tuebingen.mpg.de/mpi-is-software/o80/older/). Then, for example:
 
 ```bash
 wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_ubuntu20.04_py3.8_1.0.tar.gz
@@ -22,13 +18,11 @@ sudo make install
 sudo ldconfig
 ```
 
-## from source
+## From source
 
-### from tar
+### From tar ball
 
-Select a version by visiting : visit http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/ or http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/
-
-Then, for example:
+Select a version by visiting : [http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/](http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/) or [http://people.tuebingen.mpg.de/mpi-is-software/o80/older/](http://people.tuebingen.mpg.de/mpi-is-software/o80/older/). Then, for example:
 
 ```bash
 wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_source.tar.gz
@@ -41,24 +35,23 @@ sudo make install
 sudo ldconfig
 ```
 
-
-### via colcon workspace
+### Via colcon workspace
 
 [Colcon](https://colcon.readthedocs.io/en/released/) is the built system of [ROS2](https://docs.ros.org/en/foxy/index.html).
-The instructions below will result in the setup of a colcon workspace. Possibly, the cloned packages may be added to an existing workspace.
+The instructions below will result in the setup of a colcon workspace. Possibly, if you would like to use o80 in a ROS2 project, you may copy the cloned packages to an existing workspace.
 
 #### Adding your ssh key to github
 
 See: [github documentation](https://help.github.com/en/github/authenticating-to-github/connecting-to-github-with-ssh).
 
-All the instructions below assumes your ssh key has been activatd, i.e.:
+All the following instructions assume your ssh key has been activated, i.e.:
 
 ```bash
 # replace id_rsa by your key file name
 ssh-add ~/.ssh/id_rsa
 ```
 
-#### installing the dependencies:
+#### Installing the dependencies:
 
 ```bash
 apt install -y python3-pip cmake libcereal-dev \
@@ -74,7 +67,7 @@ pip3 install colcon-common-extensions treep \
 	     catkin-pkg sphinx breathe
 ```
 
-#### cloning the repositories
+#### Cloning the repositories
 
 Creating a folder and cloning the treep configuration:
 
@@ -90,7 +83,7 @@ Cloning all the required repositories:
 treep --clone O80
 ```
 
-#### compilation
+#### Compilation
 
 ```bash
 cd /path/to/Software
@@ -100,7 +93,7 @@ colcon build
 
 This will result in a "install" folder containing the compiled binaries
 
[…]
