### 2021-03-24 #2752 @vincentberenz
@gkthiruvathukal @traversaro @vissarion

The software is ready for the review to resume. Apologies again for this unexpected (long !) delay. The continuous integration has been fully refactored and now generates binaries and source tar balls. Here is the new documentation page:

http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/index.html

the installation page has been updated:

http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html
