+++ doc/03.integration.md
@@ -25,19 +25,19 @@ A Standalone is templated over IN and OUT (the templates of the Driver). It is a
 
 While IN is the input to a robot driver, State represents the desired state of *one* actuator of the robot.  For example, the State can be the position of a joint. 
 
-State is a user developed class which should inheritate [o80::State](https://github.com/intelligent-soft-robots/o80/blob/vberenz/doc/include/o80/state.hpp).
+State is a user developed class which should inheritate [o80::State](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/state.hpp).
 
-Standalone is templated over State because o80 allows to send desired states command to each actuator independantly. The Standalone will receive commands related to desired States, and will convert them to instance of IN before forwarding them to the driver. The user implemented Standalone class will need to implement a [convert method](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/standalone.hpp#L155) for converting instances of State into instances of IN.
+Standalone is templated over State because o80 allows to send desired states command to each actuator independantly. The Standalone will receive commands related to desired States, and will convert them to instance of IN before forwarding them to the driver. The user implemented Standalone class will need to implement a [convert method](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/standalone.hpp#L126) for converting instances of State into instances of IN.
 
-A State class does not only declare the desired state of a joint, but also [methods](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/state.hpp#L46) defining how desired states interpolates.
+A State class does not only declare the desired state of a joint, but also [methods](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/state.hpp#L43) defining how desired states interpolates.
 
- This will allow for example o80 user API to support duration command method, which will request the desired state of an actuator to reach, starting from its current desired state, a target desired state of a specified duration (this API is presented in a later section, but [here](https://github.com/intelligent-soft-robots/o80_example/blob/master/demos/duration_commands.py#L24) a preview). 
+ This will allow for example o80 user API to support duration command method, which will request the desired state of an actuator to reach, starting from its current desired state, a target desired state of a specified duration (this API is presented in a later section, but [here](https://github.com/intelligent-soft-robots/o80_example/blob/master/demos/duration_commands.py#L30) a preview). 
 
 The o80::State base class implements simple [linear interpolation methods](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/state.hxx) for basic types (int, double, float). If the user State encapsulate more complex data, and/or if the interpolation should not be linear, the "intermediate_state" methods should be overridden. 
 
 If an actuator state consists of a boolean, [o80::BoolState](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/bool_state.hpp)  can be used.
 
-If the actuator state is empty (i.e. the hardware is a sensor that takes no input), [o80::VoidState](https://github.com/intelligent-soft-robots/o80/blob/vberenz/doc/include/o80/void_state.hpp) can be used.
+If the actuator state is empty (i.e. the hardware is a sensor that takes no input), [o80::VoidState](https://github.com/intelligent-soft-robots/o80/blob/master/include/o80/void_state.hpp) can be used.
 
 For example, the toy Joint class is a valid State class:
 
@@ -56,7 +56,7 @@ A o80 [Observation](https://github.com/intelligent-soft-robots/o80/blob/master/i
 
[…]
