+++ doc/02.installation.md
@@ -32,7 +32,7 @@ Then, for example:
 
 ```bash
 wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_source.tar.gz
-tar -zxvf ./o80_ubuntu20.04_py3.8_1.0.tar.gz
+tar -zxvf ./o80_source.tar.gz
 sudo ./apt_dependencies
 sudo ./pip3_dependencies
 ./configure
