### 2021-04-07 2fb5b5a
added frontend.backend_is_active method
