### 2021-04-12 9d2cf55
added is_active method to backend
