+++ include/o80/front_end.hpp
@@ -201,6 +201,24 @@ public:
     Observation<NB_ACTUATORS, ROBOT_STATE, EXTENDED_STATE> read(
         long int iteration = -1);
 
+public:
+
+  /*! returns the time series of commands shared between the
+   *   frontend and the backend*/
+  static auto get_introspection_commands(std::string segment_id);
+
+  /*! returns the time series of (completed) command ids  
+   *  shared between the frontend and the backend*/
+  static auto get_introspection_completed_commands(std::string segment_id);
+
+  /*! returns the time series of command ids the frontend
+   *  waits completion of */
+  static auto get_introspection_waiting_for_completion(std::string segment_id);
+
+  /*! returns the time series of command ids the frontend
+   *  processed reports of completion */
+  static auto get_introspection_completion_reported(std::string segment_id);
+  
 private:
     void size_check();
     time_series::Index last_index_read_by_backend();
@@ -241,6 +259,16 @@ private:
     // for shared commands to be completed)
     CompletedCommandsTimeSeries completed_commands_;
 
+    // everytime the frontend will wait for the completion of a
+    // command (pulse_and_wait method), it will write the corresponding id in
+    // this time series. For debug and introspection.
+    CompletedCommandsTimeSeries waiting_for_completion_;
+
+    // everytime the frontend will process the information that a
+    // command has been completed by the backend, its id will be 
+    // written in this time series. For debug and introspection.
+    CompletedCommandsTimeSeries completion_reported_;
+  
     // in burst mode: used to the send the activating signal to the backend.
     LeaderPtr leader_;
 
