+++ internal/o80_internal/controllers_manager.hpp
@@ -53,6 +53,18 @@ private:
     States<NB_ACTUATORS, STATE> previous_desired_states_;
     std::array<bool, NB_ACTUATORS> initialized_;
     long int relative_iteration_;
+
+    // everytime the backend reads a new command from the 
+    // shared memory, it will write in this time series its
+    // command id. For debug and introspection.
+    CompletedCommandsTimeSeries received_commands_;
+
+    // everytime the backend starts execution of a command,
+    // it will write in this time series its
+    // command id. For debug and introspection.
+    CompletedCommandsTimeSeries starting_commands_;
+
+  
 };
 }  // namespace o80
 
