+++ include/o80/introspector.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
+#include <chrono>
 #include <mutex>
 #include <atomic>
 #include <memory>
 #include <real_time_tools/thread.hpp>
 #include "o80/front_end.hpp"
-
 
 namespace o80
 {
@@ -20,6 +20,10 @@ namespace o80
     THREAD_FUNCTION_RETURN_TYPE run_waiting_for_completion_helper(void *arg);
     template <class ROBOT_STATE>
     THREAD_FUNCTION_RETURN_TYPE run_completion_reported_helper(void *arg);
+    template <class ROBOT_STATE>
+    THREAD_FUNCTION_RETURN_TYPE run_received_helper(void *arg);
+    template <class ROBOT_STATE>
+    THREAD_FUNCTION_RETURN_TYPE run_starting_helper(void *arg);
   }
 
   template<class ROBOT_STATE>
@@ -39,12 +43,17 @@ namespace o80
     friend THREAD_FUNCTION_RETURN_TYPE internal::run_completed_commands_helper<ROBOT_STATE>(void *arg);
     friend THREAD_FUNCTION_RETURN_TYPE internal::run_waiting_for_completion_helper<ROBOT_STATE>(void *arg);
     friend THREAD_FUNCTION_RETURN_TYPE internal::run_completion_reported_helper<ROBOT_STATE>(void *arg);
+    friend THREAD_FUNCTION_RETURN_TYPE internal::run_received_helper<ROBOT_STATE>(void *arg);
+    friend THREAD_FUNCTION_RETURN_TYPE internal::run_starting_helper<ROBOT_STATE>(void *arg);
+    template<class T>
+    void run(std::shared_ptr<T> time_series,
+	     std::string prefix);
     void run_commands();
     void run_completed_commands();
     void run_waiting_for_completion();
     void run_completion_reported();
-    void run_command_ids(std::shared_ptr<CompletedCommandsTimeSeries> time_series,
-			 std::string prefix);
+    void run_received();
+    void run_starting();
   private:
     std::mutex mutex_;
     std::atomic<bool> running_;
@@ -52,10 +61,14 @@ namespace o80
     real_time_tools::RealTimeThread completed_commands_thread_;
     real_time_tools::RealTimeThread waiting_for_completion_thread_;
     real_time_tools::RealTimeThread completion_reported_thread_;
+    real_time_tools::RealTimeThread received_thread_;
+    real_time_tools::RealTimeThread starting_thread_;
     std::shared_ptr<CommandsTimeSeries> commands_;
     std::shared_ptr<CompletedCommandsTimeSeries> completed_commands_;
     std::shared_ptr<CompletedCommandsTimeSeries> waiting_for_completion_;
     std::shared_ptr<CompletedCommandsTimeSeries> completion_reported_;
+    std::shared_ptr<CompletedCommandsTimeSeries> received_;
+    std::shared_ptr<CompletedCommandsTimeSeries> starting_;
 
   private:
     static std::shared_ptr<Introspector<ROBOT_STATE>> instance_;
