### 2021-04-26 82d540d
fixed typo in installation documentation
