+++ doc/02.installation.md
@@ -27,8 +27,8 @@ Select a version by visiting : [http://people.tuebingen.mpg.de/mpi-is-software/o
 ```bash
 wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_source.tar.gz
 tar -zxvf ./o80_source.tar.gz
-sudo ./apt_dependencies
-sudo ./pip3_dependencies
+sudo ./apt-dependencies
+sudo ./pip3-dependencies
 ./configure
 make
 sudo make install
