### 2021-04-27 bf25023
added purge function
