+++ internal/o80_internal/controller.hpp
@@ -38,9 +38,8 @@ public:
     void set_completed_commands(
         CompletedCommandsTimeSeries& completed_commands);
 
-    void set_starting_commands(
-        CompletedCommandsTimeSeries& starting_commands);
-  
+    void set_starting_commands(CompletedCommandsTimeSeries& starting_commands);
+
     void set_command(const Command<STATE>& command);
 
     bool stop_current(const STATE& current_state,
@@ -49,9 +48,9 @@ public:
     void stop_all(const STATE& current_state, Microseconds control_iteration);
 
     int running(const STATE& current_state, Microseconds control_iteration);
-  
-  void purge();
-  
+
+    void purge();
+
     int size() const;
 
     const STATE& get_desired_state(long int current_iteration,
@@ -78,8 +77,8 @@ private:
 
 private:
     static std::mutex mutex_;
-  CompletedCommandsTimeSeries* completed_commands_;
-  CompletedCommandsTimeSeries* starting_commands_;
+    CompletedCommandsTimeSeries* completed_commands_;
+    CompletedCommandsTimeSeries* starting_commands_;
     std::queue<Command<STATE>> queue_;
     Command<STATE> current_command_;
     STATE desired_state_;
