+++ include/o80/introspector.hpp
@@ -1,60 +1,67 @@
 #pragma once
 
-#include <chrono>
-#include <mutex>
 #include <atomic>
+#include <chrono>
 #include <memory>
+#include <mutex>
 #include <real_time_tools/thread.hpp>
 #include "o80/front_end.hpp"
 
 namespace o80
 {
+namespace internal
+{
+template <class ROBOT_STATE>
+THREAD_FUNCTION_RETURN_TYPE run_commands_helper(void *arg);
+template <class ROBOT_STATE>
+THREAD_FUNCTION_RETURN_TYPE run_completed_commands_helper(void *arg);
+template <class ROBOT_STATE>
+THREAD_FUNCTION_RETURN_TYPE run_waiting_for_completion_helper(void *arg);
+template <class ROBOT_STATE>
+THREAD_FUNCTION_RETURN_TYPE run_completion_reported_helper(void *arg);
+template <class ROBOT_STATE>
+THREAD_FUNCTION_RETURN_TYPE run_received_helper(void *arg);
+template <class ROBOT_STATE>
+THREAD_FUNCTION_RETURN_TYPE run_starting_helper(void *arg);
+}  // namespace internal
 
-  namespace internal
-  {
-    template <class ROBOT_STATE>
-    THREAD_FUNCTION_RETURN_TYPE run_commands_helper(void *arg);
-    template <class ROBOT_STATE>
-    THREAD_FUNCTION_RETURN_TYPE run_completed_commands_helper(void *arg);
-    template <class ROBOT_STATE>
-    THREAD_FUNCTION_RETURN_TYPE run_waiting_for_completion_helper(void *arg);
-    template <class ROBOT_STATE>
-    THREAD_FUNCTION_RETURN_TYPE run_completion_reported_helper(void *arg);
-    template <class ROBOT_STATE>
-    THREAD_FUNCTION_RETURN_TYPE run_received_helper(void *arg);
-    template <class ROBOT_STATE>
-    THREAD_FUNCTION_RETURN_TYPE run_starting_helper(void *arg);
-  }
-
-  template<class ROBOT_STATE>
-  class Introspector
-  {
-  public:
+template <class ROBOT_STATE>
+class Introspector
+{
+public:
     typedef time_series::MultiprocessTimeSeries<Command<ROBOT_STATE>>
-    CommandsTimeSeries;
+        CommandsTimeSeries;
     typedef time_series::MultiprocessTimeSeries<int>
         CompletedCommandsTimeSeries;
-  public:
+
+public:
     Introspector(std::string segment_id);
     void start();
     void stop();
-  private:
-    friend THREAD_FUNCTION_RETURN_TYPE internal::run_commands_helper<ROBOT_STATE>(void *arg);
-    friend THREAD_FUNCTION_RETURN_TYPE internal::run_completed_commands_helper<ROBOT_STATE>(void *arg);
-    friend THREAD_FUNCTION_RETURN_TYPE internal::run_waiting_for_completion_helper<ROBOT_STATE>(void *arg);
-    friend THREAD_FUNCTION_RETURN_TYPE internal::run_completion_reported_helper<ROBOT_STATE>(void *arg);
-    friend THREAD_FUNCTION_RETURN_TYPE internal::run_received_helper<ROBOT_STATE>(void *arg);
-    friend THREAD_FUNCTION_RETURN_TYPE internal::run_starting_helper<ROBOT_STATE>(void *arg);
-    template<class T>
-    void run(std::shared_ptr<T> time_series,
-	     std::string prefix);
+
+private:
+    friend THREAD_FUNCTION_RETURN_TYPE
+    internal::run_commands_helper<ROBOT_STATE>(void *arg);
+    friend THREAD_FUNCTION_RETURN_TYPE
[…]
