### 2021-04-27 847db1a
added purge as a backend method
