### 2021-04-28 intelligent-soft-robots/o80#5 @vissarion
Contribution guidelines
I cannot find guidelines for new contributors and users.

It would be benefitial for the project if clear guidelines for third parties exist that address the following
- Contribute to the software
- Report issues or problems with the software
- Seek support
