### 2021-04-28 intelligent-soft-robots/o80#6 @vissarion
JOSS paper
Some comments on the paper that might improve the writing:

- related work: how this software compares to other commonly-used packages
- in Modular Implementation section you may want to include references or links to the listed packages.
- instead of citing the whole boost c++ libraries you may want to cite only Boost Interprocess (are you using another boost library?)
