### 2021-04-28 intelligent-soft-robots/o80#7 @vissarion
Functionality and tests
I managed to successfully install `o80` from the [documentation](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html).

However, I have trouble running any example. The documentation seems scattered on this aspect. What is a minimum example that can be run after a successful installation of `o80`? Does the user need to install something else? Also the automated tests that used to appear in the README are not there anymore and I cannot find them in other places too.
