### 2021-04-28 #2752 @vissarion
Hi, I did another review pass after the revision and left a few comments as issues in the target repository:

- https://github.com/intelligent-soft-robots/o80/issues/5
- https://github.com/intelligent-soft-robots/o80/issues/6
- https://github.com/intelligent-soft-robots/o80/issues/7
