### 2021-04-28 46d56ac
added initial_states method to backend
