### 2021-04-28 c5240fa
added initial_states method to frontend
