### 2021-04-28 b02d563
added frontend add_reinit_command method
