+++ README.md
@@ -1,3 +1,9 @@
+[![build](https://raw.githubusercontent.com/MPI-IS-BambooAgent/sw_badges/master/badges/plans/o80/build.svg?sanitize=true)](url)
+[![documentation](https://raw.githubusercontent.com/MPI-IS-BambooAgent/sw_badges/master/badges/plans/o80/doc.svg?sanitize=true)](url)
+[![unit_tests](https://raw.githubusercontent.com/MPI-IS-BambooAgent/sw_badges/master/badges/plans/o80/unit%20tests.svg?sanitize=true)](url)
+
+
+
 # o80
 Synchronization of c++ processes via a custom python API.
 
