+++ demos/demo_burster.cpp
@@ -1,21 +1,20 @@
-#include <iostream>
 #include <unistd.h>
+#include <iostream>
 #include "o80/burster.hpp"
-
 
 void run()
 {
-  std::string segment_id{"o80_demo_burster"};
-  o80::Burster b(segment_id);
+    std::string segment_id{"o80_demo_burster"};
+    o80::Burster b(segment_id);
 
-  while(true)
+    while (true)
     {
-      b.pulse();
-      std::cout << "pulse !" << std::endl;
+        b.pulse();
+        std::cout << "pulse !" << std::endl;
     }
 }
 
 int main()
 {
-  run();
+    run();
 }
