+++ include/o80/burster.hpp
@@ -58,43 +58,26 @@ private:
     std::shared_ptr<synchronizer::Follower> follower_;
 };
 
+/*! Client of Burster, i.e. used for commanding
+ *  an instance of Burster of the same segment_id
+ *  to perform a serie of pulses. Used internally
+ *  by Frontend's instances.
+ */
 class BursterClient
 {
 public:
-  typedef std::shared_ptr<synchronizer::Leader> LeaderPtr;
+    typedef std::shared_ptr<synchronizer::Leader> LeaderPtr;
 
-  BursterClient(std::string segment_id)
-    : leader_{segment_id+"_synchronizer",true},
-      segment_id_{segment_id}
-  {
-    set_bursting(1);
-  }
+    BursterClient(std::string segment_id);
+    void burst(int nb_iterations);
+    void final_burst();
 
-  void burst(int nb_iterations)
-  {
-    set_bursting(nb_iterations);
-    leader_.pulse();
-  }
-
-  void final_burst()
-  {
-    leader_.stop_sync();
-  }
-  
 private:
-  void set_bursting(int nb_iterations)
-  {
-    shared_memory::set<long int>(segment_id_, "bursting", nb_iterations);
-    shared_memory::set<long int>(segment_id_, "bursting_sync", nb_iterations);
-  }
+    void set_bursting(int nb_iterations);
+
 private:
-  synchronizer::Leader leader_;
-  std::string segment_id_;
-
-
-
-  
+    synchronizer::Leader leader_;
+    std::string segment_id_;
 };
 
-  
 }  // namespace o80
