### 2021-05-04 669c734
update paper: compare to actionlib, link to core-robotics, link only to boost interprocess and link to HYSR/pam doc
