+++ paper/paper.bib
@@ -3,7 +3,8 @@
  title = {The Boost C++ Libraries},
  year = {2011},
  isbn = {0982219199},
- publisher = {XML Press}
+ publisher = {XML Press},
+ url = {https://www.boost.org/doc/libs/1_63_0/doc/html/interprocess.html}
 }
 
 @misc{cereal,
@@ -11,6 +12,12 @@
  year = {2017},
  url = {http://uscilab.github.io/cereal/},
  title = {cereal - A C++11 library for serialization}
+}
+
+@misc{corerobotics,
+author= {Naveau Maximilien and Berenz Vincent and Widemaier Felix and Viereck Julian and W{\"u}thrich Manuel},
+year = {2020},
+url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
 }
 
 @misc{pybind11,
@@ -86,4 +93,10 @@
  eprint={1606.01540},
  archivePrefix={arXiv},
  primaryClass={cs.LG}
+}
+
+@misc{pamsource,
+author= {Berenz Vincent and Simon Guist and Dieter Buechler},
+year = {2020},
+url = {http://people.tuebingen.mpg.de/mpi-is-software/pam/docs/pam_documentation/index.html}
 }
