+++ paper/paper.bib
@@ -8,27 +8,33 @@
 }
 
 @misc{cereal,
- author = {Grant W. Shane and Voorhies Randolph},
+ author = {Shane Grant W. and Randolph Voorhies},
  year = {2017},
  url = {http://uscilab.github.io/cereal/},
  title = {cereal - A C++11 library for serialization}
 }
 
 @misc{corerobotics,
-author= {Naveau Maximilien and Berenz Vincent and Widemaier Felix and Viereck Julian and W{\"u}thrich Manuel},
+author= {Maximilien Naveau and Vincent Berenz and Felix Widemaier and Julian Viereck and Manuel W{\"u}thrich},
 year = {2020},
 url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
 }
 
+@misc{actionlib,
+author= {Michael Carroll and Jakob Perron and Eitan Marder-Eppstein and Vijay Pradeep  and Mikael Arguedas}
+year = {2009},
+url = {http://wiki.ros.org/actionlib}
+}
+
 @misc{pybind11,
- author = {Wenzel Jakob and Jason Rhinelander and Dean Moldovan},
+ author = {Jakob Wenzel and Rhinelander Jason and Moldovan Dean},
  year = {2016},
  url = {https://github.com/pybind/pybind11},
  title = {pybind11 — Seamless operability between C++11 and Python}
 } 
 
 @misc{sl,
- author = {Stefan Schaal},
+ author = {Stefan, Schaal},
  year = {2009},
  url = {http://www-clmc.usc.edu/publications/S/schaal-TRSL.pdf},
  title = {The SL simulation and real-time control software package}
@@ -50,7 +56,7 @@ url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
 
 @misc{icsds,
  title={Identifying Causal Structure in Dynamical Systems},
- author={Dominik Baumann and Friedrich Solowjow and Karl H. Johansson and Sebastian Trimpe},
+ author={Baumann Dominik and Solowjow Friedrich and Johansson Karl H. and Trimpe Sebastian},
  year={2020},
  eprint={2006.03906},
  archivePrefix={arXiv},
@@ -59,7 +65,7 @@ url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
 
 @misc{pam,
   title={Learning to Play Table Tennis From Scratch using Muscular Robots},
-  author={Dieter Büchler and Simon Guist and Roberto Calandra and Vincent Berenz and Bernhard Schölkopf and Jan Peters},
+  author={Büchler Dieter and Guist Simon and Calandra Roberto and Berenz Vincent and Schölkopf Bernhard and Peters Jan},
   year={2020},
   eprint={2006.05935},
   archivePrefix={arXiv},
@@ -88,7 +94,7 @@ url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
 
 @misc{gym,
  title={OpenAI Gym},
- author={Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
+ author={Brockman Greg and Cheung Vicki and Pettersson Ludwig and Schneider Jonas and Schulman John and Tang Jie and Zaremba Wojciech},
  year={2016},
  eprint={1606.01540},
  archivePrefix={arXiv},
@@ -96,7 +102,7 @@ url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
 }
 
 @misc{pamsource,
-author= {Berenz Vincent and Simon Guist and Dieter Buechler},
[…]
