### 2021-05-04 cf316cf
paper: fixed typos in statement of need section
