+++ paper/paper.bib
@@ -21,7 +21,7 @@ url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
 }
 
 @misc{actionlib,
-author= {Michael Carroll and Jakob Perron and Eitan Marder-Eppstein and Vijay Pradeep  and Mikael Arguedas}
+author= {Michael Carroll and Jakob Perron and Eitan Marder-Eppstein and Vijay Pradeep and Mikael Arguedas},
 year = {2009},
 url = {http://wiki.ros.org/actionlib}
 }
