+++ paper/paper.bib
@@ -102,7 +102,7 @@ url = {http://wiki.ros.org/actionlib}
 }
 
 @misc{pamsource,
-author= {Berenz Vincent and Guist Simon and Buechler Dieter},
+author= {Vincent Berenz and Simon Guist and Dieter Büchler},
 year = {2020},
 url = {http://people.tuebingen.mpg.de/mpi-is-software/pam/docs/pam_documentation/index.html}
 }
