+++ doc/03.support.md
@@ -1,0 +1,9 @@
+# Support
+
+## mail
+
+For support and question, please contact [Vincent Berenz](https://is.mpg.de/person/vberenz).
+
+## issues
+
+If you have ideas for improvement, or would like to report a bug, you may create issues on [github](https://github.com/intelligent-soft-robots/o80).
