+++ doc/02.installation.md
@@ -111,3 +111,9 @@ In a python3 terminal:
 ```python
 import o80
 ```
+
+## Running the demos
+
+The [o80_example](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/index.html) package provides
+usage examples of o80 running on a dummy robot. 
+
