+++ doc/09.developers.md
@@ -1,0 +1,272 @@
+# Developers
+
+This guide is here for those who would like to upgrade/extend o80, or one of its dependencies.
+
+For updating the code, the colcon workspace installation is required, as described [here](file:///home/vberenz/Workspaces/pam_colcon/workspace/install/o80/share/o80/docs/sphinx/html/doc/02.installation.html#via-colcon-workspace).
+
+The workflow is based on these tools: github, treep, ament, colcon, pybind11 and gtest. This page will give a brief usage guide for them, before describing the guidelines to follow.
+
+
+## Tools
+
+### Github
+
+The source repository is hosted in the "Intelligent Soft Robots" [domain](https://github.com/intelligent-soft-robots) of github.
+As treep (see below) uses ssh, you need to register your ssh key to github (see [here](https://docs.github.com/en/github/authenticating-to-github/adding-a-new-ssh-key-to-your-github-account))
+
+### Treep
+
+To get o80 compiled and running, several git repositories must be cloned.
+As it would be annoying to clone them one by one, we use the project manager [treep](https://pypi.org/project/treep/), which allows to clone several repositories at once.
+
+#### installation
+
+```bash
+pip install treep
+```
+
+#### configuration repository
+
+treep needs a configuration folder. We use [treep_isr](https://github.com/intelligent-soft-robots/treep_isr.git):
+
+```bash
+mkdir Software # or another folder name
+cd Software
+git clone https://github.com/intelligent-soft-robots/treep_isr.git
+```
+
+In the folder treep_isr, you will find the file *repositories.yaml* which provides repositories names, url and default branch (see also configuration.yaml). Once treep_isr has been cloned, you may see for example the list of repositories:
+
+```bash
+treep # treep without argument prints some help
+treep --repos
+```
+
+Projects (i.e. repositories grouped together) are described in *projects.yaml*.
+To see the list of projects:
+
+```bash
+treep --projects
+```
+
+To see information about a specific projects:
+
+```bash
+treep --project O80
+```
+
+#### cloning
+
+You can use treep to clone a project (i.e. cloning all the repositories of the project):
+
+```bash
+treep --clone O80
+```
+
+#### status
+
+The status argument is very useful to see the (git) status of each cloned repos (i.e. if they are behind origin, beyond origin, have modified files, etc)
+
+```bash
+treep --status
+```
+
+#### pull 
+
+You can use treep to update all the repositories
+
+```bash
+treep --pull
+```
+
+
+#### calling from anywhere
+
+Note that you do not need to be in the Software folder to call treep, you may call treep for any of its subfolder, e.g.
+
+```bash
+cd Software/workspace/src/pam_interface
+treep --status # still provides the status of all repositories.
+```
+
+### Ament packages
+
+(Almost) all repositories contain the code for an ament package. An ament package is a "extended" CMake package, i.e. ament provides some cmake functionalities meant to make developers life easier. The documentation can be found [here](https://docs.ros.org/en/foxy/Guides/Ament-CMake-Documentation.html).
+
+An ament package contains typically:
+
+- C++ code, in the include/package_name and the source folders
+
+- python wrappers over this C++ code in the srcpy folder
+
+- native python code in the python folder
+
+- unit tests in the tests folder
+
+- configuration files
+
+- a demos folder, containing source code providing example of usage of the code provided by the package
+
+- a package.xml file listing dependencies
+
+- a CMakeLists.txt file containing the CMake commands required for compilation.
+
[…]
