### 2021-05-04 intelligent-soft-robots/o80#6 @vincentberenz
@vissarion Thanks for the comments. I made the following updates to the papers:

- in the first section, I added a short comparison to ROS's actionlib, which would be the most similar software used in robotics.
- I added a link to the documentation to the packages mentioned in the "Modular Implementation" section.
- I replaced the link to Boost to the link of Boost's interprocess library

I took the opportunity to make these other updates:

- I updated my affiliation, which changed recently
- I fixed the bib file, in which the author names was not properly given ("firstname lastname" instead of "lastname firstname")
- I added a reference to the source code for the HYSR software (last section)
