### 2021-05-04 intelligent-soft-robots/o80#5 @vincentberenz
@vissarion I made the following updates to the documentation:

- added a support section: http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/03.support.html
- added a developers section:http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/09.developers.html
