+++ README.md
@@ -13,4 +13,5 @@ See : the [documentation](http://people.tuebingen.mpg.de/mpi-is-software/o80/doc
 # Author
 Vincent Berenz, Max Planck Institute for Intelligent Systems, Empirical Inference Department
 
-<sup><sub>dev note: the unit tests for o80 are with the [o80_example]((https://github.com/intelligent-soft-robots/o80_example)) package
+# Unit-tests
+The unit tests for o80 are with the [o80_example]((https://github.com/intelligent-soft-robots/o80_example)) package
