+++ README.md
@@ -14,4 +14,4 @@ See : the [documentation](http://people.tuebingen.mpg.de/mpi-is-software/o80/doc
 Vincent Berenz, Max Planck Institute for Intelligent Systems, Empirical Inference Department
 
 # Unit-tests
-The unit tests for o80 are with the [o80_example]((https://github.com/intelligent-soft-robots/o80_example)) package
+The unit tests for o80 are with the [o80_example](https://github.com/intelligent-soft-robots/o80_example) package
