### 2021-05-04 intelligent-soft-robots/o80#7 @vincentberenz
o80 is a heavily templated library, and an unfortunate consequence of it is that even the simplest example becomes quite involved.

The "minimal" example for the usage of o80 is in this package: https://github.com/intelligent-soft-robots/o80_example .

The introduction of the documentation points to this example package, but indeed it was not emphasized enough.

I added this to the documentation:
http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html#running-the-demos

The unit-tests for o80 are in o80_example (because o80 is templated, concrete implementations are required, and this is what o80_example provides). I made this now clear in the readme of o80:
https://github.com/intelligent-soft-robots/o80
