+++ paper/paper.bib
@@ -27,7 +27,7 @@ url = {http://wiki.ros.org/actionlib}
 }
 
 @misc{pybind11,
- author = {Jakob Wenzel and Rhinelander Jason and Moldovan Dean},
+ author = {Jakob Wenzel and Jason Rhinelander and Dean Moldovan},
  year = {2016},
  url = {https://github.com/pybind/pybind11},
  title = {pybind11 — Seamless operability between C++11 and Python}
@@ -56,7 +56,7 @@ url = {http://wiki.ros.org/actionlib}
 
 @misc{icsds,
  title={Identifying Causal Structure in Dynamical Systems},
- author={Baumann Dominik and Solowjow Friedrich and Johansson Karl H. and Trimpe Sebastian},
+ author={Dominik Baumann and Friedrich Solowjow and Karl H. Johansson and Sebastian Trimpe},
  year={2020},
  eprint={2006.03906},
  archivePrefix={arXiv},
@@ -65,7 +65,7 @@ url = {http://wiki.ros.org/actionlib}
 
 @misc{pam,
   title={Learning to Play Table Tennis From Scratch using Muscular Robots},
-  author={Büchler Dieter and Guist Simon and Calandra Roberto and Berenz Vincent and Schölkopf Bernhard and Peters Jan},
+  author={Dieter Büchler and Simon Guist and Roberto Calandra and Berenz Vincent and Bernhard Schölkopf and Jan Peters},
   year={2020},
   eprint={2006.05935},
   archivePrefix={arXiv},
@@ -94,7 +94,7 @@ url = {http://wiki.ros.org/actionlib}
 
 @misc{gym,
  title={OpenAI Gym},
- author={Brockman Greg and Cheung Vicki and Pettersson Ludwig and Schneider Jonas and Schulman John and Tang Jie and Zaremba Wojciech},
+ author={Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
  year={2016},
  eprint={1606.01540},
  archivePrefix={arXiv},
