+++ paper/paper.bib
@@ -34,7 +34,7 @@ url = {http://wiki.ros.org/actionlib}
 } 
 
 @misc{sl,
- author = {Stefan, Schaal},
+ author = {Schaal, Stefan},
  year = {2009},
  url = {http://www-clmc.usc.edu/publications/S/schaal-TRSL.pdf},
  title = {The SL simulation and real-time control software package}
