### 2021-05-19 24a508e
removing affiliation subscripts as everybody with same affiliation
