### 2021-05-19 fc5f434
in paper, writing explicit affiliation index, as otherwise JOSS's whedon has pdf compilation issue
