### 2021-05-19 #2752 @vincentberenz
I made a minor change in the paper: I updated the affiliations so that they align with some new guidelines of our institute
