### 2021-06-01 intelligent-soft-robots/o80#8 @traversaro
JOSS paper comments (traversaro comments)
Hi @vincentberenz,
  I read the latest version of the paper in https://github.com/openjournals/joss-reviews/issues/2752#issuecomment-844095526 .

Some (minor) comments (I quote the original part of the paper:

> These packages are available on the GitHub domains “intelligent soft robots,” “machines in motion,” “mpi-is” and “open dynamic robot initiative.”

What do you mean by "GiHub domains"? The typical term for organizational accounts in GitHub is "GitHub Organizations", see for example https://docs.github.com/en/organizations/collaborating-with-groups-in-organizations/about-organizations . Furthermore, it may be easier to use exactly the name of the organization, so `intelligent-soft-robots`, `machines-in-motion` and so on so forth.

> o80 is in some aspects similar to ROS’s actionlib (Carroll et al., 2009), as both allow a process to create and monitor execution of commands executed by another process, with python and c++ interoperability

Nitpick: c++/C++

> The complete list, the sources, the binaries as well as the documentation of theses packages70can be found online (Naveau et al., 2020)

The "(Naveau et al., 2020)" reference links to http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/, that is a directory listing, without any explanation of what this directory contains. Without a landing page, probably it is better to just link http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/docs/ , that at least contains a list of the packages.
