### 2021-06-02 intelligent-soft-robots/o80#2 @traversaro
Thanks @vincentberenz and sorry for the delay! The link in the previous comment is now broken.
