### 2021-06-02 intelligent-soft-robots/o80#2 @traversaro
I am testing the new installation procedure documented in http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html#from-binaries, and at least in the latest archive, it seems that the scripts are named `apt-dependencies` and `pip3-dependencies`, instead of `apt_dependencies` and `pip3_dependencies` as listed in the docs.
