### 2021-06-02 intelligent-soft-robots/o80#9 @traversaro
Problem encountered trying to run demos
Part of https://github.com/openjournals/joss-reviews/issues/2752 .

Thanks a lot @vincentberenz for the nice installation procedure (I added some minor comments in https://github.com/intelligent-soft-robots/o80/issues/2#issuecomment-852779972). I tried to follow the "Installation from binaries", but I found the following problems trying the demos. Note that I installed the `o80_ubuntu18.04_py3.6_1.34.tar.gz` tarball on an headless Ubuntu 18.04 machine.

* Minor: The "Running the demos" section in http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html#running-the-demos links to http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/index.html . If I follow the link http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/index.html, I find a list of step, that brings me back to http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html that is the page from where I started. Perhaps it could make sense to link directly to http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/doc/02_demos.html ?

* Once I reached http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/doc/02_demos.html, I found the indication "downloading the example repository" without any additional indication, so i downloaded the default branch. However, if I try to run the first example, it fails with:
~~~
root@4442114f7f41:~/o80/o80_example/demos# python3 ./standalone_backend.py
Traceback (most recent call last):
  File "./standalone_backend.py", line 4, in <module>
    import signal_handler
ModuleNotFoundError: No module named 'signal_handler'
~~~
I am not sure what the problem is, but I imagine it is related to the fact that the `master` branch was not the correct version of o80 to use, but I could not find any reference to a `v1.34` version (see https://github.com/openjournals/joss-reviews/issues/2752#issuecomment-709391077 for a related comment). Furthermore, even the `master` branch version did not seem to be coherent with the documentation in the website, as no `standalone_backend_bursting.py` nor `bursting.py` files can be found in the `demos` .

 * By switching `o80_examples` to the `v1.0` tag, I was able to avoid the `import signal_handler` problem, related to the fact that I was running everything on an headless machine:
 ~~~
 root@4442114f7f41:~/o80/o80_example/demos# python3 ./standalone_backend.py
Warning this thread is not going to be real time.
qt.qpa.xcb: could not connect to display
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: eglfs, linuxfb, minimal, minimalegl, offscreen, vnc, wayland-egl, wayland, wayland-xcomposite-egl, wayland-xcomposite-glx, webgl, xcb.

Aborted
~~~
I will now setup a machine with X support, but I think that it could be helpful to mention in the requirements that o80 must be run on a machine with Graphical\X support, not an headless one, early in o80 documentation, for example in http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html#installation .
