+++ paper/paper.bib
@@ -17,7 +17,7 @@
 @misc{corerobotics,
 author= {Maximilien Naveau and Vincent Berenz and Felix Widemaier and Julian Viereck and Manuel W{\"u}thrich},
 year = {2020},
-url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/}
+url = {http://people.tuebingen.mpg.de/mpi-is-software/corerobotics/docs}
 }
 
 @misc{actionlib,
