### 2021-06-02 18ea749
fixed documentation: apt_dependencies-> apt-dependencies (same for pip3 dependencies)
