+++ doc/02.installation.md
@@ -11,8 +11,8 @@ To install, first select a version by visiting : [http://people.tuebingen.mpg.de
 ```bash
 wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_ubuntu20.04_py3.8_1.0.tar.gz
 tar -zxvf ./o80_ubuntu20.04_py3.8_1.0.tar.gz
-sudo ./apt_dependencies
-sudo ./pip3_dependencies
+sudo ./apt-dependencies
+sudo ./pip3-dependencies
 ./configure
 sudo make install
 sudo ldconfig
