### 2021-06-02 2018bcf
doc: installation redirecting at the end to the demos page of o80_example
