+++ doc/02.installation.md
@@ -114,6 +114,5 @@ import o80
 
 ## Running the demos
 
-The [o80_example](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/index.html) package provides
-usage examples of o80 running on a dummy robot. 
+The o80_example package provides usage examples of o80 running on a dummy robot, see [here](http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80_example/doc/02_demos.html).
 
