### 2021-06-02 intelligent-soft-robots/o80#8 @vincentberenz
@traversaro Thanks a lot for finding the issues ! I made all the advised changes
