### 2021-06-02 intelligent-soft-robots/o80#2 @vincentberenz
@traversaro thanks for catching these typos. I fixed them
