### 2021-06-02 intelligent-soft-robots/o80#9 @vincentberenz
@traversaro  Thanks a lot for installing and running the demos !

- I fixed the circular links issues in the documentation, using the proposed solution
- The master branch of o80_example should have worked. I will try to replicate the issue, and once found it, update the continuous integration, which should have caught it (but did not)
- I updated the install page of o80_example in the doc to inform X support is required (it is required only for the plotting of the demos)

I will ping this issue again once I made the updates. I understand there are some confusion regarding versioning, I will also answer them in the other issue.
