### 2021-06-02 intelligent-soft-robots/o80#10 @vincentberenz
@traversaro

Thanks a lot for having taking the time to discuss this. Even if these points are not part of the "official checklist", I am grateful I can address them during this review.

The versioning is currently confusing because everytime an update over any dependencies of o80 is updated, the continuous integration upload new tar balls if the compilation and tests are successful. So currently, even updating a typo in the documentation triggers a new "version". I must improve this.

I believe the issues related to installation a mostly a lack of documentation.
Before calling "make" or "make install", the configure script is called and set the installation path. It is possible for user to set then user paths, but this is not documented. Also: if an virtual environment is active, then by default the python libraries will be installed in it (also not documented).
For the pip3 dependencies installation, it also seem to me that if a virtual environment is active, it will install for it (but I will double check, and possibly document).

I will look at these points carefully and make updates. I will not be available the next two weeks, so this may take a bit. Thanks for your patience and for your suggestions.
