### 2021-06-02 intelligent-soft-robots/o80#10 @traversaro
> I believe the issues related to installation a mostly a lack of documentation.
Before calling "make" or "make install", the configure script is called and set the installation path. It is possible for user to set then user paths, but this is not documented. Also: if an virtual environment is active, then by default the python libraries will be installed in it (also not documented).
For the pip3 dependencies installation, it also seem to me that if a virtual environment is active, it will install for it (but I will double check, and possibly document).

Great, I imagine that, good to know that this is the case.
