+++ include/o80/state6d.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "o80/statexd.hpp"
+
+namespace o80
+{
+typedef StateXd<double, double, double, double, double, double> State6d;
+}  // namespace o80
