### 2021-06-25 0557041
documentation: updated installation instructions
