+++ doc/02.installation.md
@@ -12,7 +12,16 @@ To install, first select a version by visiting : [http://people.tuebingen.mpg.de
 wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_ubuntu20.04_py3.8_1.0.tar.gz
 tar -zxvf ./o80_ubuntu20.04_py3.8_1.0.tar.gz
 sudo ./apt-dependencies
-sudo ./pip3-dependencies
+# for global installation
+#sudo ./pip3-dependencies
+# for user installation or currently activated virtual environment installation
+./pip3-dependencies
+# configuration. By default, this will
+# install libraries and includes to /usr/local
+# and python packages to the active python (including currently
+# activated virtual environment, if any). To overwrite the default,
+# for example: 
+# ./configure --prefix=/usr/ --pythondir=/usr/local/lib/python3.8/dist-packages 
 ./configure
 sudo make install
 sudo ldconfig
@@ -28,7 +37,16 @@ Select a version by visiting : [http://people.tuebingen.mpg.de/mpi-is-software/o
 wget http://people.tuebingen.mpg.de/mpi-is-software/o80/latest/o80_source.tar.gz
 tar -zxvf ./o80_source.tar.gz
 sudo ./apt-dependencies
-sudo ./pip3-dependencies
+# for global installation
+#sudo ./pip3-dependencies
+# for user installation or currently activated virtual environment installation
+./pip3-dependencies
+# configuration. By default, this will
+# install libraries and includes to /usr/local
+# and python packages to the active python (including currently
+# activated virtual environment, if any). To overwrite the default,
+# for example: 
+# ./configure --prefix=/usr/ --pythondir=/usr/local/lib/python3.8/dist-packages 
 ./configure
 make
 sudo make install
