### 2021-06-25 intelligent-soft-robots/o80#10 @vincentberenz
@traversaro

**Installation**

I updated the document regarding the installation, to point out how to change the installation prefix, or target user installation (or virtual environment installation) for python (in comments of the sections "From binaries" and "From source"):

http://people.tuebingen.mpg.de/mpi-is-software/o80/docs/o80/doc/02.installation.html

****versioning****

Until recently, the download page was cluttered with a lots of version, with no log of the modifs. This was due to the fact that the continuous integration was uploading new tar files at each successful run, and runs would occur for any minor change (e.g. fix of a typo in the documentation of one of the dependency repo, etc).

So, I introduced this file:
https://github.com/intelligent-soft-robots/o80/blob/master/version.txt

and the continuous integration will continue to run for every tiny change, but will upload only if the version number has been changed.

There is currently only 1 version available, but I will introduce a file logging the updates introduced for each version.
