### 2021-06-28 #2752 @gkthiruvathukal
@vissarion and @traversaro Can you give me a sense of whether @vincentberenz latest updates are sufficient? The earlier correspondence suggests we are good to go. I'd like to move to accept this paper by end of week if there are no additional issues to address.
