### 2021-06-28 #2752 @vincentberenz
@gkthiruvathukal

Thanks for pinging !
I still have this issue to address:

https://github.com/intelligent-soft-robots/o80/issues/9
