### 2021-06-28 intelligent-soft-robots/o80#10 @traversaro
Great @vincentberenz, thanks a lot!
