### 2021-06-29 intelligent-soft-robots/o80#9 @vincentberenz
@traversaro

Thanks for your patience !

The import issue was fixed, and hopefully you will no longer encounter issues when running the demos.
(the issue was that CMake was installing some of the packages for python2.7 instead of python3, hence the import issues).
