### 2021-08-04 #2752 @vincentberenz
@traversaro @vissarion thanks again for your time and your useful comments

@gkthiruvathukal

A robotic group of the University of Aachen  just finished developing o80 drivers for the Franka robots (which are currently quite popular in academia). I would like to update the paper with a reference to their repo. They told me it would make public very soon. As soon as this happens, I would update the paper accordingly (minimal change, smthg like "o80 drivers for the Franka robots are also availalbe [reference]"), and then go through the todo list you provided.
