### 2021-08-04 #2752 @gkthiruvathukal
@vincentberenz This is absolutely fine. Please signal us when you have made the updates and have completed the checklist. The change you are proposing seems routine, so we do not need to do a full re-review by any means. Just make sure the article PDF generates properly!
