+++ python/o80/__init__.py
@@ -1,0 +1,1 @@
+from o80_wrp import *
