+++ CMakeLists.txt
@@ -92,6 +92,7 @@ target_include_directories(demo_burster
 target_link_libraries(demo_burster ${PROJECT_NAME})
 set(all_targets ${all_targets} demo_burster)
 
+
 ###################
 # Python wrappers #
 ###################
@@ -101,23 +102,23 @@ target_link_libraries(${PROJECT_NAME}_py PRIVATE pybind11::module)
 target_link_libraries(${PROJECT_NAME}_py PRIVATE ${PYTHON_LIBRARIES})
 target_link_libraries(${PROJECT_NAME}_py PRIVATE ${PROJECT_NAME})
 set_target_properties(${PROJECT_NAME}_py
-  PROPERTIES PREFIX "" SUFFIX "${PYTHON_MODULE_EXTENSION}"
-  OUTPUT_NAME ${PROJECT_NAME})
+  PROPERTIES PREFIX "" SUFFIX "${PYTHON_MODULE_EXTENSION}")
+set_target_properties(${PROJECT_NAME}_py PROPERTIES
+  OUTPUT_NAME ${PROJECT_NAME}_wrp)
 target_include_directories(
   ${PROJECT_NAME}_py
   PUBLIC $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
          $<INSTALL_INTERFACE:include> SYSTEM
-  PUBLIC ${PYTHON_INCLUDE_DIRS})
+	 PUBLIC ${PYTHON_INCLUDE_DIRS})
 _ament_cmake_python_get_python_install_dir()
 install(TARGETS ${PROJECT_NAME}_py DESTINATION ${PYTHON_INSTALL_DIR})
 
 
-#######################
-# debian control file #
-#######################
+######################
+# Python Native code #
+######################
 
-# debian control file in
-# the o80_example package
+ament_python_install_package(${PROJECT_NAME} PACKAGE_DIR python/${PROJECT_NAME})
 
 
 #################
