+++ paper/paper.bib
@@ -7,6 +7,13 @@
  url = {https://www.boost.org/doc/libs/1_63_0/doc/html/interprocess.html}
 }
 
+@misc{o80_franka,
+ author = {Kyrylo, Sovailo},
+ title = {o80 Franka},
+ year = {2021},
+ url = {https://github.com/Data-Science-in-Mechanical-Engineering/franka_o80}
+}
+
 @misc{cereal,
  author = {Shane Grant W. and Randolph Voorhies},
  year = {2017},
