### 2021-09-17 #2752 @vincentberenz
@gkthiruvathukal

The o80's drivers for the Franka robot have been made open source, and I updated the paper to refer to them
