### 2021-10-04 #2752 @gkthiruvathukal
@vincentberenz, So there are no changes to the code based on your point about the o80 drivers, right?
Thanks for your prompt response.
