### 2021-10-04 #2752 @vincentberenz
@gkthiruvathukal I confirm the only change is the update of the paper (referring to the franka panda drivers). None of the code has been changed.
