### 2021-10-04 #2752 @vincentberenz
> @vincentberenz So I can just use your v1.0 tag for the release, right?

yes
