+++ paper/paper.md
@@ -35,20 +35,20 @@ bibliography: paper.bib
 
 # Statement of need
 
-o80 (pronounced "oh-eighty") is a software for synchronizing and organizing message exchange between (realtime) processes via simple customized Python APIs. Its target domain is robotics and machine learning. Our motivation for developing o80 is to ease the setup of robotics experiments (i.e. integration of various hardware and software) by machine learning scientists. Such setup typically requires time and technical effort, especially when realtime processes are involved. Ideally, scientists should have access to a simple Python API that hides the lower level communication details and simply allows to send actions and receive observations. o80 is a tool box for creating such API.
+o80 (pronounced "oh-eighty") is software for synchronizing and organizing message exchange between (realtime) processes via simple customized Python APIs. Its target domain is robotics and machine learning. Our motivation for developing o80 is to ease the setup of robotics experiments (i.e., integration of various hardware and software) by machine learning scientists. Such setup typically requires time and technical effort, especially when realtime processes are involved. Ideally, scientists should have access to a simple Python API that hides the lower level communication details and simply allows the sending of actions and receiving of observations. o80 is a tool box for creating such API.
 
-o80 is in some aspects similar to ROS's actionlib [@actionlib], as both allow a process to create and monitor execution of commands executed by another process, with python and C++ interoperability. A particularity of o80 is its support for queues of command and the possibilities to request the server to automatically perform linear interpolation between them. o80 also introduces new synchronization methods (see "bursting mode" in the next section). Contrary to actionlib, o80 does not support network communication as it expects the processes it orchestrates to run on the same computer.
+o80 is, in some aspects, similar to ROS's actionlib [@actionlib], as both allow a process to create and monitor execution of commands executed by another process, with Python and C++ interoperability. A particular feature of o80 is its support for queues of command and the ability to request the server to automatically perform linear interpolation between them. o80 also introduces new synchronization methods (see "bursting mode" in the next section). Unlike actionlib, o80 does not support network communication as it expects the processes it orchestrates to run on the same computer.
 
 # Overview
 
 For implementing synchronization, o80 organizes two types of processes:
 
-- A server process encapsulates an instance of o80 back-end and an instance of a driver. At each iteration, the back-end instance computes for each actuator the desired state to be transmitted to the hardware via the driver. The back-end also reads sensory information from the driver. Typically, the server process is programmed in C++ with consideration for realtime.
-- A client process encapsulates instances of o80 front-end, which provides: 1) an interface to send commands to the back-end (i.e. requests to compute desired states), 2) methods for querying sensory information, and 3) methods for synchronizing the client with the server process.
+- A server process encapsulates an instance of the o80 back-end and an instance of a driver. At each iteration, the back-end instance computes for each actuator the desired state to be transmitted to the hardware via the driver. The back-end also reads sensory information from the driver. Typically, the server process is programmed in C++ in consideration of realtime needs.
+- A client process encapsulates instances of the o80 front-end, which provides: 1) an interface to send commands to the back-end (e.g., requests to compute desired states), 2) methods for querying sensory information, and 3) methods for synchronizing the client with the server process.
 
 In the background, back-end and front-end(s) communicate by exchanging serialized object instances via a interprocess shared memory. Serialization is based on the cereal library [@cereal], and the shared memory is based on the Boost interprocess library [@boost].
 
-o80 is templated over actuator states and driver; and may therefore support a wide range of systems.
+o80 is templated over actuator states and driver, and may therefore support a wide range of systems.
 
 o80's core API supports:
 
@@ -61,34 +61,34 @@ o80's core API supports:
 
[…]
