### 2021-10-04 #2752 @danielskatz
👋 @vincentberenz - I've suggested some minor changes in the paper in https://github.com/intelligent-soft-robots/o80/pull/11 - please merge these or let me know what you disagree with, then we can proceed to acceptance and publication.
